Add explicit float literal type
Separate from 32-bit float type - as DXC does treat literals as separate entities now
UntypedInt types are now IntLiteral so they line up nicely with other enum variants
Replace placeholder "long" support with placeholder uint64_t / int64_t support

+++ hlsl/tests/byte_buffer.hlsl
@@ -21,10 +21,10 @@ void test() {
     g_output.Store2(0u, (uint2)x2);
     g_output.Store3(0u, (uint3)x3);
     g_output.Store4(0u, (uint4)x4);
-    g_output.Store((uint)0.0, (uint)0.0);
-    g_output.Store2((uint)0.0, (uint2)0.0);
-    g_output.Store3((uint)0.0, (uint3)0.0);
-    g_output.Store4((uint)0.0, (uint4)0.0);
+    g_output.Store(0u, 0u);
+    g_output.Store2(0u, (uint2)0.0);
+    g_output.Store3(0u, (uint3)0.0);
+    g_output.Store4(0u, (uint4)0.0);
     uint outInt;
     g_input.GetDimensions(outInt);
     g_output.GetDimensions(outInt);
